refactor: move ActionFormDialog + CategoryFormDialog to components/shared/forms/
Both dialogs were used by Templates importing across feature folders --
the pattern docs/ARCHITECTURE.md §5 explicitly forbids. New shared/forms/
barrel keeps the architectural invariant clean.
