+++ src/admin-app/src/components/shared/forms/index.js
@@ -1,0 +1,8 @@
+/**
+ * @file index.js
+ * @summary Barrel for shared form-dialog components used across multiple
+ * features (e.g. Challenges and Templates both manage actions and categories).
+ */
+
+export { default as ActionFormDialog } from "./ActionFormDialog";
+export { default as CategoryFormDialog } from "./CategoryFormDialog";
